services: refactor CLI (#1089)
* service/auth: refactor cli

* service/broker: tidy up Run function

* service/config: refactor cli

* service/debug: refactor cli

* service/health: tidy up Run function

* service/network: refactor cli

* service/registry: remove redundant cli commands

* service/router: tidy up Run function

* service/runtime: refactor cli

* service/store: refactor cli

* service/store: tidy up Run function

* tests: update to use 'micro services' command

* test: update to handle errors

* fix merge conflict (missing import)

* Fix test

* test: fix store tests

* Tweak
