+++ service/health/health.go
@@ -9,56 +9,48 @@ import (
 	"github.com/micro/go-micro/v2"
 	"github.com/micro/go-micro/v2/client"
 	proto "github.com/micro/go-micro/v2/debug/service/proto"
-	log "github.com/micro/go-micro/v2/logger"
+	"github.com/micro/go-micro/v2/logger"
 	mcli "github.com/micro/micro/v2/client/cli"
 	qcli "github.com/micro/micro/v2/internal/command/cli"
 	"golang.org/x/net/context"
 )
 
-var (
-	healthAddress = ":8088"
-	serverAddress string
-	serverName    string
+const healthAddress = ":8088"
 
-	// Flags specific to the health service
-	Flags = []cli.Flag{
-		&cli.StringFlag{
-			Name:    "check_service",
-			Usage:   "Name of the service to query",
-			EnvVars: []string{"MICRO_HEALTH_CHECK_SERVICE"},
-		},
-		&cli.StringFlag{
-			Name:    "check_address",
-			Usage:   "Set the service address to query",
-			EnvVars: []string{"MICRO_HEALTH_CHECK_ADDRESS"},
-		},
-	}
-)
+// Flags specific to the health service
+var Flags = []cli.Flag{
+	&cli.StringFlag{
+		Name:    "check_service",
+		Usage:   "Name of the service to query",
+		EnvVars: []string{"MICRO_HEALTH_CHECK_SERVICE"},
+	},
+	&cli.StringFlag{
+		Name:    "check_address",
+		Usage:   "Set the service address to query",
+		EnvVars: []string{"MICRO_HEALTH_CHECK_ADDRESS"},
+	},
+}
 
+// Run micro health
 func Run(ctx *cli.Context, srvOpts ...micro.Option) {
-	log.Init(log.WithFields(map[string]interface{}{"service": "health"}))
-
 	// just check service health
 	if ctx.Args().Len() > 0 {
 		mcli.Print(qcli.QueryHealth)(ctx)
 		return
 	}
 
-	serverName = ctx.String("check_service")
-	serverAddress = ctx.String("check_address")
+	srvOpts = append([]micro.Option{
+		micro.Address(healthAddress),
+	}, srvOpts...)
 
-	if addr := ctx.String("address"); len(addr) > 0 {
-		healthAddress = addr
-	}
+	serverName := ctx.String("check_service")
+	serverAddress := ctx.String("check_address")
 
-	if len(healthAddress) == 0 {
-		log.Fatal("health address not set")
-	}
 	if len(serverName) == 0 {
-		log.Fatal("service name not set")
+		logger.Fatal("service name not set")
 	}
 	if len(serverAddress) == 0 {
-		log.Fatal("service address not set")
+		logger.Fatal("service address not set")
 	}
 
 	http.HandleFunc("/health", func(w http.ResponseWriter, r *http.Request) {
@@ -75,10 +67,10 @@ func Run(ctx *cli.Context, srvOpts ...micro.Option) {
 		fmt.Fprint(w, "OK")
 	})
 
-	log.Infof("Health check running at %s/health", healthAddress)
-	log.Infof("Health check defined for %s at %s", serverName, serverAddress)
+	logger.Infof("Health check running at %s/health", healthAddress)
+	logger.Infof("Health check defined for %s at %s", serverName, serverAddress)
 
 	if err := http.ListenAndServe(healthAddress, nil); err != nil {
-		log.Fatal(err)
+		logger.Fatal(err)
 	}
 }
