+++ service/runtime/runtime.go
@@ -16,10 +16,11 @@ import (
 )
 
 var (
-	// Name of the runtime
-	Name = "go.micro.runtime"
-	// Address of the runtime
-	Address = ":8088"
+	// name of the runtime
+	name = "go.micro.runtime"
+	// address of the runtime
+	address = ":8088"
+
 	// Flags specific to the runtime service
 	Flags = []cli.Flag{
 		&cli.StringFlag{
@@ -42,8 +43,6 @@ var (
 
 // Run the runtime service
 func Run(ctx *cli.Context, srvOpts ...micro.Option) {
-	log.Init(log.WithFields(map[string]interface{}{"service": "runtime"}))
-
 	// Get the profile
 	var prof []string
 	switch ctx.String("profile") {
@@ -60,15 +59,15 @@ func Run(ctx *cli.Context, srvOpts ...micro.Option) {
 	}
 
 	if len(ctx.String("address")) > 0 {
-		Address = ctx.String("address")
+		address = ctx.String("address")
 	}
 
 	if len(ctx.String("server_name")) > 0 {
-		Name = ctx.String("server_name")
+		name = ctx.String("server_name")
 	}
 
-	if len(Address) > 0 {
-		srvOpts = append(srvOpts, micro.Address(Address))
+	if len(address) > 0 {
+		srvOpts = append(srvOpts, micro.Address(address))
 	}
 
 	// create runtime
@@ -78,7 +77,7 @@ func Run(ctx *cli.Context, srvOpts ...micro.Option) {
 	}
 
 	// append name
-	srvOpts = append(srvOpts, micro.Name(Name))
+	srvOpts = append(srvOpts, micro.Name(name))
 
 	// new service
 	service := micro.NewService(srvOpts...)
@@ -115,106 +114,4 @@ func Run(ctx *cli.Context, srvOpts ...micro.Option) {
 		log.Errorf("failed to stop: %s", err)
 		os.Exit(1)
 	}
-}
-
-// flags is shared flags so we don't have to continually re-add
-func flags() []cli.Flag {
-	return []cli.Flag{
-		&cli.StringFlag{
-			Name:  "source",
-			Usage: "Set the source url of the service e.g github.com/micro/services",
-		},
-		&cli.StringFlag{
-			Name:  "image",
-			Usage: "Set the image to use for the container",
-		},
-		&cli.StringFlag{
-			Name:  "command",
-			Usage: "Command to exec",
-		},
-		&cli.StringFlag{
-			Name:  "args",
-			Usage: "Command args",
-		},
-		&cli.StringFlag{
-			Name:  "type",
-			Usage: "The type of service operate on",
-		},
-		&cli.StringSliceFlag{
-			Name:  "env_vars",
-			Usage: "Set the environment variables e.g. foo=bar",
-		},
-	}
-}
-
-func Commands(options ...micro.Option) []*cli.Command {
-	command := []*cli.Command{
-		{
-			// In future we'll also have `micro run [x]` hence `micro run service` requiring "service"
-			Name:  "run",
-			Usage: RunUsage,
-			Description: `Examples:
-			micro run github.com/micro/examples/helloworld
-			micro run .  # deploy local folder to your local micro server
-			micro run ../path/to/folder # deploy local folder to your local micro server
-			micro run helloworld # deploy latest version, translates to micro run github.com/micro/services/helloworld
-			micro run helloworld@9342934e6180 # deploy certain version
-			micro run helloworld@branchname	# deploy certain branch`,
-			Flags: flags(),
-			Action: func(ctx *cli.Context) error {
-				runService(ctx, options...)
-				return nil
-			},
-		},
-		{
-			Name:  "update",
-			Usage: UpdateUsage,
-			Description: `Examples:
-			micro update github.com/micro/examples/helloworld
-			micro update .  # deploy local folder to your local micro server
-			micro update ../path/to/folder # deploy local folder to your local micro server
-			micro update helloworld # deploy master branch, translates to micro update github.com/micro/services/helloworld
-			micro update helloworld@branchname	# deploy certain branch`,
-			Flags: flags(),
-			Action: func(ctx *cli.Context) error {
-				updateService(ctx, options...)
-				return nil
-			},
-		},
-		{
-			Name:  "kill",
-			Usage: KillUsage,
-			Flags: flags(),
-			Description: `Examples:
-			micro kill github.com/micro/examples/helloworld
-			micro kill .  # kill service deployed from local folder
-			micro kill ../path/to/folder # kill service deployed from local folder
-			micro kill helloworld # kill serviced deployed from master branch, translates to micro kill github.com/micro/services/helloworld
-			micro kill helloworld@branchname	# kill service deployed from certain branch`,
-			Action: func(ctx *cli.Context) error {
-				killService(ctx, options...)
-				return nil
-			},
-		},
-		{
-			Name:  "status",
-			Usage: GetUsage,
-			Flags: flags(),
-			Action: func(ctx *cli.Context) error {
-				getService(ctx, options...)
-				return nil
-			},
-		},
-		{
-			Name:  "logs",
-			Usage: "Get logs for a service",
-			Flags: logFlags(),
-			Action: func(ctx *cli.Context) error {
-				getLogs(ctx, options...)
-				return nil
-			},
-		},
-	}
-
-	return command
 }
