+++ service/store/cli/databases.go
@@ -11,8 +11,8 @@ import (
 	inclient "github.com/micro/micro/v2/internal/client"
 )
 
-// Databases is the entrypoint for micro store databases
-func Databases(ctx *cli.Context) error {
+// databases is the entrypoint for micro store databases
+func databases(ctx *cli.Context) error {
 	client := inclient.New(ctx)
 	dbReq := client.NewRequest(ctx.String("store"), "Store.Databases", &storeproto.DatabasesRequest{})
 	dbRsp := &storeproto.DatabasesResponse{}
@@ -25,8 +25,8 @@ func Databases(ctx *cli.Context) error {
 	return nil
 }
 
-// Tables is the entrypoint for micro store tables
-func Tables(ctx *cli.Context) error {
+// tables is the entrypoint for micro store tables
+func tables(ctx *cli.Context) error {
 	ns, err := namespace.Get(util.GetEnv(ctx).Name)
 	if err != nil {
 		return err
