+++ service/store/store.go
@@ -5,33 +5,29 @@ import (
 	"github.com/micro/go-micro/v2"
 	log "github.com/micro/go-micro/v2/logger"
 	pb "github.com/micro/go-micro/v2/store/service/proto"
-	mcli "github.com/micro/micro/v2/client/cli"
-	"github.com/micro/micro/v2/internal/helper"
 	"github.com/micro/micro/v2/service/store/handler"
 )
 
 var (
-	// Name of the store service
-	Name = "go.micro.store"
-	// Address is the store address
-	Address = ":8002"
+	// name of the store service
+	name = "go.micro.store"
+	// address is the store address
+	address = ":8002"
 )
 
-// Run runs the micro server
+// Run micro store
 func Run(ctx *cli.Context, srvOpts ...micro.Option) {
-	log.Init(log.WithFields(map[string]interface{}{"service": "store"}))
-
 	if len(ctx.String("server_name")) > 0 {
-		Name = ctx.String("server_name")
+		name = ctx.String("server_name")
 	}
 	if len(ctx.String("address")) > 0 {
-		Address = ctx.String("address")
+		address = ctx.String("address")
 	}
 
 	// Initialise service
 	service := micro.NewService(
-		micro.Name(Name),
-		micro.Address(Address),
+		micro.Name(name),
+		micro.Address(address),
 	)
 
 	// the store handler
@@ -42,15 +38,4 @@ func Run(ctx *cli.Context, srvOpts ...micro.Option) {
 	if err := service.Run(); err != nil {
 		log.Fatal(err)
 	}
-}
-
-// Commands is the cli interface for the store service
-func Commands(options ...micro.Option) []*cli.Command {
-	command := &cli.Command{
-		Name:        "store",
-		Action:      helper.UnexpectedSubcommand,
-		Subcommands: mcli.StoreCommands(),
-	}
-
-	return []*cli.Command{command}
 }
